feat: enhancements to Git config

+++ scripts/git.sh
@@ -4,11 +4,17 @@
 git config --global init.defaultBranch main
 
 git config --global core.ignorecase true
+git config --global core.editor nano
 
 git config --global pull.rebase true
 
 git config --global rerere.enabled true
 git config --global rerere.autoupdate true
 
-cp -f ./git/.gitignore_global ~/.gitignore
-git config --global core.excludesfile ~/.gitignore
+# Use interactive-rebase-tool (https://github.com/MitMaro/git-interactive-rebase-tool, installed separately via Brew).
+git config --global sequence.editor interactive-rebase-tool
+
+if [[ ! -f ~/.gitignore ]]; then
+  cp -f ./git/.gitignore_global ~/.gitignore
+  git config --global core.excludesfile ~/.gitignore
+fi
